TrikeGo/requirements.txt
Django==5.2.6
python-dotenv>=1.0.0
supabase>=1.8.3
psycopg2-binary>=2.9
djangorestframework>=3.14.0
channels>=4.0.0
channels-redis>=4.1.0
openrouteservice>=2.3.3
requests>=2.31.0
gunicorn==23.0.0
whitenoise==6.8.2
dj_database_url
django-redis>=5.2.0
celery>=5.3.0
pywebpush>=1.14.0